Introduce a new IBatchTextFormatter for handling batches of log events
This is useful if you want to format log events inside a table of an html email for example.

This new interface allows you to write a header, e.g. `<html><body><table>` before writing a batch of log events and a footer, e.g. `</table></body></html>`after the batch of log events is written.